Finished Main features, improved Layout, applied Bugfixes.
Implemented Search feature making it possible to search food in the
current fridge.
Improved layout of food activity making it more intuitive and user
friendly.  Also made main activitys scrollable and tiltable.
Fixed many bugs appearing while registering and the opening activitys
for the first time.
Added a feature that checks if  the chosen username is not used yet.